test : SessionRepositoryTest에서 Student와 관련된 부분을 StudentRepositoryTest로 옮긴다
